Conflicts:
	src/org/broad/igv/sam/AlignmentTrack.java
	src/org/broad/igv/sam/SamAlignment.java
	src/org/broad/igv/ui/PreferencesEditor.java

+++ test/src/org/broad/igv/util/TestUtils.java
@@ -1,19 +1,12 @@
 /*
- * Copyright (c) 2007-2011 by The Broad Institute of MIT and Harvard.  All Rights Reserved.
+ * Copyright (c) 2007-2012 The Broad Institute, Inc.
+ * SOFTWARE COPYRIGHT NOTICE
+ * This software and its documentation are the copyright of the Broad Institute, Inc. All rights are reserved.
+ *
+ * This software is supplied without any warranty or guaranteed support whatsoever. The Broad Institute is not responsible for its use, misuse, or functionality.
  *
  * This software is licensed under the terms of the GNU Lesser General Public License (LGPL),
  * Version 2.1 which is available at http://www.opensource.org/licenses/lgpl-2.1.php.
- *
- * THE SOFTWARE IS PROVIDED "AS IS." THE BROAD AND MIT MAKE NO REPRESENTATIONS OR
- * WARRANTES OF ANY KIND CONCERNING THE SOFTWARE, EXPRESS OR IMPLIED, INCLUDING,
- * WITHOUT LIMITATION, WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE, NONINFRINGEMENT, OR THE ABSENCE OF LATENT OR OTHER DEFECTS, WHETHER
- * OR NOT DISCOVERABLE.  IN NO EVENT SHALL THE BROAD OR MIT, OR THEIR RESPECTIVE
- * TRUSTEES, DIRECTORS, OFFICERS, EMPLOYEES, AND AFFILIATES BE LIABLE FOR ANY DAMAGES
- * OF ANY KIND, INCLUDING, WITHOUT LIMITATION, INCIDENTAL OR CONSEQUENTIAL DAMAGES,
- * ECONOMIC DAMAGES OR INJURY TO PROPERTY AND LOST PROFITS, REGARDLESS OF WHETHER
- * THE BROAD OR MIT SHALL BE ADVISED, SHALL HAVE OTHER REASON TO KNOW, OR IN FACT
- * SHALL KNOW OF THE POSSIBILITY OF THE FOREGOING.
  */
 
 package org.broad.igv.util;
@@ -31,6 +24,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.io.File;
 import java.io.IOException;
+import java.util.concurrent.Future;
 
 /**
  * @author jrobinso
@@ -38,13 +32,13 @@ import java.io.IOException;
  */
 @Ignore
 public class TestUtils {
-    public static String DATA_DIR = "test/data";
+    public static String DATA_DIR = "test/data/";
     public static String dataFileName = DATA_DIR + "/genomes/hg18.unittest.genome";
     public static String AVAILABLE_FTP_URL = "ftp://ftp.broadinstitute.org/pub/igv/TEST/test.txt";
     public static String UNAVAILABLE_FTP_URL = "ftp://www.example.com/file.txt";
     //This is so ant can set the large data directory
     private static String LARGE_DATA_DIR_KEY = "LARGE_DATA_DIR";
-    public static String LARGE_DATA_DIR = "test/largedata";
+    public static String LARGE_DATA_DIR = "test/largedata/";
 
     static {
         LARGE_DATA_DIR = System.getProperty(LARGE_DATA_DIR_KEY, LARGE_DATA_DIR);
@@ -56,6 +50,12 @@ public class TestUtils {
         Globals.READ_TIMEOUT = 60 * 1000;
         Globals.CONNECT_TIMEOUT = 60 * 1000;
         FTPClient.READ_TIMEOUT = 60 * 1000;
+
+        //Create output directory if it doesn't exist
+        File outDir = new File(DATA_DIR, "out");
+        if (!outDir.exists()) {
+            outDir.mkdir();
+        }
     }
 
     public static void setUpHeadless() {
@@ -104,6 +104,7 @@ public class TestUtils {
         }
         org.junit.Assume.assumeTrue(!headless);
         setUpTestEnvironment();
+        Globals.setHeadless(false);
         IGV igv;
         //If IGV is already open, we get the instance.
         try {
@@ -121,6 +122,18 @@ public class TestUtils {
             igv.loadGenome(genomeFile, null);
         }
         return igv;
+    }
+
+    /**
+     * Loads the session into IGV. This blocks until the session
+     * is loaded.
+     *
+     * @param igv
+     * @param sessionPath
+     * @throws InterruptedException
+     */
+    public static void loadSession(IGV igv, String sessionPath) throws InterruptedException {
+        igv.doRestoreSession(sessionPath, null, false);
     }
 
     public static boolean checkHeadlessEnvironment() {
@@ -172,7 +185,10 @@ public class TestUtils {
         if (outputDir.isDirectory()) {
             File[] listFiles = outputDir.listFiles();
             for (File fi : listFiles) {
-                fi.delete();
+                //Keep hidden files and directories
+                if (!fi.getName().startsWith(".")) {
+                    fi.delete();
+                }
             }
         }
     }
